imp - brk|doc - Defer to delegates
---

For true experience, we've studied using delegates for native methods, and saw that it was useful for scenarios where a .NET Standard 2.0 library needs to refer to multiple native libraries. For this very reason, we've decided to replace the current method of loading native libraries - which doesn't work well because it involves copying the native library files to the main executable path, which creates duplicates that we don't need, and, in Linux, loading them using dlopen, which has no effect to the CoreCLR because subsequent calls to the p/invoke methods will try to search the library again, which fails - with the better method that involves function delegates that are so useful.

This is a breaking change, because you'll have to replace the entire p/invoke method list with the delegate function prototypes. Create a delegate for each symbol you found in the library and use GetNativeMethodDelegate(), pointing it to your delegate. Then, call your delegate.

---

Type: imp
Breaking: True
Doc Required: True
Backport Required: False
Part: 1/1